[Driver] Move argparser to driver
