+++ diana/byoc/utils.py
@@ -680,64 +680,3 @@ def get_gdb_output(gdb_log_path="debug/gdb.txt"):
     # Values are returned from GDB in one big one-dimensional tensor
     # Reshaping here such that it matches the output
     return values_from_test
-    
-
-def parse_cli_options() -> Tuple[argparse.Namespace, str]:
-    '''
-    Utility function that reads arguments from command line
-    usage: see script_name.py -h
-    '''
-    parser = argparse.ArgumentParser(description="Utility argparser\
-                                                  for example scripts")
-    parser.add_argument('--target', dest='target',
-                        help="Target string to pass onto TVMC, note that '-device=arm_cpu' is appended to the string later",
-                        default="soma_dory, c")
-    parser.add_argument('--device', dest='device',
-                        choices = ("pulp", "x86"),
-                        help="Device to make binary for (which makefile to call), (default 'pulp')",
-                        default="pulp")
-    parser.add_argument('--profile', dest='measurement',
-                        help="Insert PULP performance counters into generated C code; for each individual kernel, for the entire TVM artefact, or don't insert performance counters (default)",
-                        choices=("individual", "global", "power", None),
-                        default=None)
-    parser.add_argument('--interactive', dest='interactive',
-                        action='store_const', const=True,
-                        help="Wait for user input to have performed measurement to parse profiler results",
-                        default=False)
-    parser.add_argument('--no-fusion', dest='fusion',
-                        help="Set TVM's Relay Fusion pass maximum fusion depth to 0",
-                        action='store_const', const=False,
-                        default=True)
-    parser.add_argument('--weight-bits', dest='weight_bits', type=int,
-                        help="Number of bits per weight. This affects the selection of the digital/analog core",
-                        choices=(8, 2),
-                        default=8)
-    parser.add_argument('--gcc-opt', dest='gcc_opt',
-                        choices = (0, 1, 2, 3), type=int,
-                        help="Set the gcc optimization level in pulprt makefile, (default Makefile.pulprt)",
-                        default=3)
-    parser.add_argument('--makefile', dest='makefile',
-            help="Set different path for pulprt makefile (default for device = x86: Makefile.x86, default for  device = pulp: Makefile.pulprt")
-    args = parser.parse_args()
-    # Some options shouldn't be used together
-    if args.device == "x86":
-        if "soma_dory" in args.target:
-            raise ValueError("Dory codegen can not be compiled for --device=\"x86\", only for --device=\"pulp\"")
-        if args.measurement is not None:
-            raise ValueError("Profiling is not available for --device=\"x86\", only for --device=\"pulp\"")
-    # Set default makefile if no other is specified
-    if args.makefile is None:
-        if args.device == "pulp":
-            args.makefile = "Makefile.pulprt"
-        elif args.device == "x86":
-            args.makefile = "Makefile.x86"
-    adapt_gcc_opt(args.makefile, args.gcc_opt)
-    # Return string which identifies options
-    def get_options_string(args: argparse.Namespace):
-        fusion_name = "fused" if args.fusion else "unfused"
-        target_name = "dory" if args.target == "soma_dory, c" else "c"
-        options_string = f"{args.device}_{target_name}_{fusion_name}" + \
-                   f"_O{args.gcc_opt}_{args.measurement}"
-        return options_string
-
-    return args, get_options_string(args)
